moved all references to BitNodeMultipliers over to calc in spot getBitNodeMultipliers
