+++ src/BitNode/BitNode.tsx
@@ -1,6 +1,4 @@
 import React from "react";
-import { BitNodeMultipliers } from "./BitNodeMultipliers";
-import { IPlayer } from "../PersonObjects/IPlayer";
 import { IMap } from "../types";
 import { FactionNames } from "../Faction/data/FactionNames";
 import { CityName } from "../Locations/data/CityNames";
@@ -579,360 +577,3 @@ BitNodes["BitNode21"] = new BitNode(21, 2, "", "COMING SOON");
 BitNodes["BitNode22"] = new BitNode(22, 2, "", "COMING SOON");
 BitNodes["BitNode23"] = new BitNode(23, 2, "", "COMING SOON");
 BitNodes["BitNode24"] = new BitNode(24, 2, "", "COMING SOON");
-
-export function initBitNodeMultipliers(p: IPlayer): void {
-  if (p.bitNodeN == null) {
-    p.bitNodeN = 1;
-  }
-  for (const mult of Object.keys(BitNodeMultipliers)) {
-    if (BitNodeMultipliers.hasOwnProperty(mult)) {
-      BitNodeMultipliers[mult] = 1;
-    }
-  }
-  // Special case.
-  BitNodeMultipliers.StaneksGiftExtraSize = 0;
-
-  switch (p.bitNodeN) {
-    case 1: // Source Genesis (every multiplier is 1)
-      break;
-    case 2: // Rise of the Underworld
-      BitNodeMultipliers.HackingLevelMultiplier = 0.8;
-      BitNodeMultipliers.ServerGrowthRate = 0.8;
-      BitNodeMultipliers.ServerMaxMoney = 0.2;
-      BitNodeMultipliers.ServerStartingMoney = 0.4;
-      BitNodeMultipliers.CrimeMoney = 3;
-      BitNodeMultipliers.InfiltrationMoney = 3;
-      BitNodeMultipliers.FactionWorkRepGain = 0.5;
-      BitNodeMultipliers.FactionPassiveRepGain = 0;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 2;
-      BitNodeMultipliers.StaneksGiftExtraSize = -6;
-      BitNodeMultipliers.PurchasedServerSoftcap = 1.3;
-      BitNodeMultipliers.CorporationSoftCap = 0.9;
-      BitNodeMultipliers.WorldDaemonDifficulty = 5;
-      break;
-    case 3: // Corporatocracy
-      BitNodeMultipliers.HackingLevelMultiplier = 0.8;
-      BitNodeMultipliers.RepToDonateToFaction = 0.5;
-      BitNodeMultipliers.AugmentationRepCost = 3;
-      BitNodeMultipliers.AugmentationMoneyCost = 3;
-      BitNodeMultipliers.ServerMaxMoney = 0.2;
-      BitNodeMultipliers.ServerStartingMoney = 0.2;
-      BitNodeMultipliers.ServerGrowthRate = 0.2;
-      BitNodeMultipliers.ScriptHackMoney = 0.2;
-      BitNodeMultipliers.CompanyWorkMoney = 0.25;
-      BitNodeMultipliers.CrimeMoney = 0.25;
-      BitNodeMultipliers.HacknetNodeMoney = 0.25;
-      BitNodeMultipliers.HomeComputerRamCost = 1.5;
-      BitNodeMultipliers.PurchasedServerCost = 2;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 0.75;
-      BitNodeMultipliers.StaneksGiftExtraSize = -2;
-      BitNodeMultipliers.PurchasedServerSoftcap = 1.3;
-      BitNodeMultipliers.GangSoftcap = 0.9;
-      BitNodeMultipliers.WorldDaemonDifficulty = 2;
-      BitNodeMultipliers.GangUniqueAugs = 0.5;
-      break;
-    case 4: // The Singularity
-      BitNodeMultipliers.ServerMaxMoney = 0.15;
-      BitNodeMultipliers.ServerStartingMoney = 0.75;
-      BitNodeMultipliers.ScriptHackMoney = 0.2;
-      BitNodeMultipliers.CompanyWorkMoney = 0.1;
-      BitNodeMultipliers.CrimeMoney = 0.2;
-      BitNodeMultipliers.HacknetNodeMoney = 0.05;
-      BitNodeMultipliers.CompanyWorkExpGain = 0.5;
-      BitNodeMultipliers.ClassGymExpGain = 0.5;
-      BitNodeMultipliers.FactionWorkExpGain = 0.5;
-      BitNodeMultipliers.HackExpGain = 0.4;
-      BitNodeMultipliers.CrimeExpGain = 0.5;
-      BitNodeMultipliers.FactionWorkRepGain = 0.75;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 1.5;
-      BitNodeMultipliers.StaneksGiftExtraSize = 0;
-      BitNodeMultipliers.PurchasedServerSoftcap = 1.2;
-      BitNodeMultipliers.WorldDaemonDifficulty = 3;
-      BitNodeMultipliers.GangUniqueAugs = 0.5;
-      break;
-    case 5: // Artificial intelligence
-      BitNodeMultipliers.ServerMaxMoney = 2;
-      BitNodeMultipliers.ServerStartingSecurity = 2;
-      BitNodeMultipliers.ServerStartingMoney = 0.5;
-      BitNodeMultipliers.ScriptHackMoney = 0.15;
-      BitNodeMultipliers.HacknetNodeMoney = 0.2;
-      BitNodeMultipliers.CrimeMoney = 0.5;
-      BitNodeMultipliers.InfiltrationRep = 1.5;
-      BitNodeMultipliers.InfiltrationMoney = 1.5;
-      BitNodeMultipliers.AugmentationMoneyCost = 2;
-      BitNodeMultipliers.HackExpGain = 0.5;
-      BitNodeMultipliers.CorporationValuation = 0.5;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 1.3;
-      BitNodeMultipliers.StaneksGiftExtraSize = 0;
-      BitNodeMultipliers.PurchasedServerSoftcap = 1.2;
-      BitNodeMultipliers.WorldDaemonDifficulty = 1.5;
-      BitNodeMultipliers.GangUniqueAugs = 0.5;
-      break;
-    case 6: // Bladeburner
-      BitNodeMultipliers.HackingLevelMultiplier = 0.35;
-      BitNodeMultipliers.ServerMaxMoney = 0.4;
-      BitNodeMultipliers.ServerStartingMoney = 0.5;
-      BitNodeMultipliers.ServerStartingSecurity = 1.5;
-      BitNodeMultipliers.ScriptHackMoney = 0.75;
-      BitNodeMultipliers.CompanyWorkMoney = 0.5;
-      BitNodeMultipliers.CrimeMoney = 0.75;
-      BitNodeMultipliers.InfiltrationMoney = 0.75;
-      BitNodeMultipliers.CorporationValuation = 0.2;
-      BitNodeMultipliers.HacknetNodeMoney = 0.2;
-      BitNodeMultipliers.HackExpGain = 0.25;
-      BitNodeMultipliers.DaedalusAugsRequirement = 1.166; // Results in 35 Augs needed
-      BitNodeMultipliers.PurchasedServerSoftcap = 2;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 0.5;
-      BitNodeMultipliers.StaneksGiftExtraSize = 2;
-      BitNodeMultipliers.GangSoftcap = 0.7;
-      BitNodeMultipliers.CorporationSoftCap = 0.9;
-      BitNodeMultipliers.WorldDaemonDifficulty = 2;
-      BitNodeMultipliers.GangUniqueAugs = 0.2;
-      break;
-    case 7: // Bladeburner 2079
-      BitNodeMultipliers.BladeburnerRank = 0.6;
-      BitNodeMultipliers.BladeburnerSkillCost = 2;
-      BitNodeMultipliers.AugmentationMoneyCost = 3;
-      BitNodeMultipliers.HackingLevelMultiplier = 0.35;
-      BitNodeMultipliers.ServerMaxMoney = 0.4;
-      BitNodeMultipliers.ServerStartingMoney = 0.5;
-      BitNodeMultipliers.ServerStartingSecurity = 1.5;
-      BitNodeMultipliers.ScriptHackMoney = 0.5;
-      BitNodeMultipliers.CompanyWorkMoney = 0.5;
-      BitNodeMultipliers.CrimeMoney = 0.75;
-      BitNodeMultipliers.InfiltrationMoney = 0.75;
-      BitNodeMultipliers.CorporationValuation = 0.2;
-      BitNodeMultipliers.HacknetNodeMoney = 0.2;
-      BitNodeMultipliers.HackExpGain = 0.25;
-      BitNodeMultipliers.FourSigmaMarketDataCost = 2;
-      BitNodeMultipliers.FourSigmaMarketDataApiCost = 2;
-      BitNodeMultipliers.DaedalusAugsRequirement = 1.166; // Results in 35 Augs needed
-      BitNodeMultipliers.PurchasedServerSoftcap = 2;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 0.9;
-      BitNodeMultipliers.StaneksGiftExtraSize = -1;
-      BitNodeMultipliers.GangSoftcap = 0.7;
-      BitNodeMultipliers.CorporationSoftCap = 0.9;
-      BitNodeMultipliers.WorldDaemonDifficulty = 2;
-      BitNodeMultipliers.GangUniqueAugs = 0.2;
-      break;
-    case 8: // Ghost of Wall Street
-      BitNodeMultipliers.ScriptHackMoney = 0.3;
-      BitNodeMultipliers.ScriptHackMoneyGain = 0;
-      BitNodeMultipliers.ManualHackMoney = 0;
-      BitNodeMultipliers.CompanyWorkMoney = 0;
-      BitNodeMultipliers.CrimeMoney = 0;
-      BitNodeMultipliers.HacknetNodeMoney = 0;
-      BitNodeMultipliers.InfiltrationMoney = 0;
-      BitNodeMultipliers.RepToDonateToFaction = 0;
-      BitNodeMultipliers.CorporationValuation = 0;
-      BitNodeMultipliers.CodingContractMoney = 0;
-      BitNodeMultipliers.StaneksGiftExtraSize = -7;
-      BitNodeMultipliers.PurchasedServerSoftcap = 4;
-      BitNodeMultipliers.GangSoftcap = 0;
-      BitNodeMultipliers.CorporationSoftCap = 0;
-      BitNodeMultipliers.GangUniqueAugs = 0;
-      break;
-    case 9: // Hacktocracy
-      BitNodeMultipliers.HackingLevelMultiplier = 0.5;
-      BitNodeMultipliers.StrengthLevelMultiplier = 0.45;
-      BitNodeMultipliers.DefenseLevelMultiplier = 0.45;
-      BitNodeMultipliers.DexterityLevelMultiplier = 0.45;
-      BitNodeMultipliers.AgilityLevelMultiplier = 0.45;
-      BitNodeMultipliers.CharismaLevelMultiplier = 0.45;
-      BitNodeMultipliers.PurchasedServerLimit = 0;
-      BitNodeMultipliers.HomeComputerRamCost = 5;
-      BitNodeMultipliers.CrimeMoney = 0.5;
-      BitNodeMultipliers.ScriptHackMoney = 0.1;
-      BitNodeMultipliers.HackExpGain = 0.05;
-      BitNodeMultipliers.ServerStartingMoney = 0.1;
-      BitNodeMultipliers.ServerMaxMoney = 0.1;
-      BitNodeMultipliers.ServerStartingSecurity = 2.5;
-      BitNodeMultipliers.CorporationValuation = 0.5;
-      BitNodeMultipliers.FourSigmaMarketDataCost = 5;
-      BitNodeMultipliers.FourSigmaMarketDataApiCost = 4;
-      BitNodeMultipliers.BladeburnerRank = 0.9;
-      BitNodeMultipliers.BladeburnerSkillCost = 1.2;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 0.5;
-      BitNodeMultipliers.StaneksGiftExtraSize = 2;
-      BitNodeMultipliers.GangSoftcap = 0.8;
-      BitNodeMultipliers.CorporationSoftCap = 0.7;
-      BitNodeMultipliers.WorldDaemonDifficulty = 2;
-      BitNodeMultipliers.GangUniqueAugs = 0.25;
-      break;
-    case 10: // Digital Carbon
-      BitNodeMultipliers.HackingLevelMultiplier = 0.35;
-      BitNodeMultipliers.StrengthLevelMultiplier = 0.4;
-      BitNodeMultipliers.DefenseLevelMultiplier = 0.4;
-      BitNodeMultipliers.DexterityLevelMultiplier = 0.4;
-      BitNodeMultipliers.AgilityLevelMultiplier = 0.4;
-      BitNodeMultipliers.CharismaLevelMultiplier = 0.4;
-      BitNodeMultipliers.CompanyWorkMoney = 0.5;
-      BitNodeMultipliers.CrimeMoney = 0.5;
-      BitNodeMultipliers.HacknetNodeMoney = 0.5;
-      BitNodeMultipliers.ManualHackMoney = 0.5;
-      BitNodeMultipliers.ScriptHackMoney = 0.5;
-      BitNodeMultipliers.CodingContractMoney = 0.5;
-      BitNodeMultipliers.InfiltrationMoney = 0.5;
-      BitNodeMultipliers.CorporationValuation = 0.5;
-      BitNodeMultipliers.AugmentationMoneyCost = 5;
-      BitNodeMultipliers.AugmentationRepCost = 2;
-      BitNodeMultipliers.HomeComputerRamCost = 1.5;
-      BitNodeMultipliers.PurchasedServerCost = 5;
-      BitNodeMultipliers.PurchasedServerLimit = 0.6;
-      BitNodeMultipliers.PurchasedServerMaxRam = 0.5;
-      BitNodeMultipliers.BladeburnerRank = 0.8;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 0.75;
-      BitNodeMultipliers.StaneksGiftExtraSize = -3;
-      BitNodeMultipliers.PurchasedServerSoftcap = 1.1;
-      BitNodeMultipliers.GangSoftcap = 0.9;
-      BitNodeMultipliers.CorporationSoftCap = 0.9;
-      BitNodeMultipliers.WorldDaemonDifficulty = 2;
-      BitNodeMultipliers.GangUniqueAugs = 0.25;
-      break;
-    case 11: //The Big Crash
-      BitNodeMultipliers.HackingLevelMultiplier = 0.6;
-      BitNodeMultipliers.HackExpGain = 0.5;
-      BitNodeMultipliers.ServerMaxMoney = 0.1;
-      BitNodeMultipliers.ServerStartingMoney = 0.1;
-      BitNodeMultipliers.ServerGrowthRate = 0.2;
-      BitNodeMultipliers.ServerWeakenRate = 2;
-      BitNodeMultipliers.CrimeMoney = 3;
-      BitNodeMultipliers.CompanyWorkMoney = 0.5;
-      BitNodeMultipliers.HacknetNodeMoney = 0.1;
-      BitNodeMultipliers.AugmentationMoneyCost = 2;
-      BitNodeMultipliers.InfiltrationMoney = 2.5;
-      BitNodeMultipliers.InfiltrationRep = 2.5;
-      BitNodeMultipliers.CorporationValuation = 0.1;
-      BitNodeMultipliers.CodingContractMoney = 0.25;
-      BitNodeMultipliers.FourSigmaMarketDataCost = 4;
-      BitNodeMultipliers.FourSigmaMarketDataApiCost = 4;
-      BitNodeMultipliers.PurchasedServerSoftcap = 2;
-      BitNodeMultipliers.CorporationSoftCap = 0.9;
-      BitNodeMultipliers.WorldDaemonDifficulty = 1.5;
-      BitNodeMultipliers.GangUniqueAugs = 0.75;
-      break;
-    case 12: {
-      //The Recursion
-      let sf12Lvl = 0;
-      for (let i = 0; i < p.sourceFiles.length; i++) {
-        if (p.sourceFiles[i].n === 12) {
-          sf12Lvl = p.sourceFiles[i].lvl;
-        }
-      }
-      const inc = Math.pow(1.02, sf12Lvl);
-      const dec = 1 / inc;
-
-      // Multiplier for number of augs needed for Daedalus increases
-      // up to a maximum of 1.34, which results in 40 Augs required
-      BitNodeMultipliers.DaedalusAugsRequirement = Math.min(inc, 1.34);
-
-      BitNodeMultipliers.HackingLevelMultiplier = dec;
-      BitNodeMultipliers.StrengthLevelMultiplier = dec;
-      BitNodeMultipliers.DefenseLevelMultiplier = dec;
-      BitNodeMultipliers.DexterityLevelMultiplier = dec;
-      BitNodeMultipliers.AgilityLevelMultiplier = dec;
-      BitNodeMultipliers.CharismaLevelMultiplier = dec;
-
-      BitNodeMultipliers.ServerMaxMoney = dec;
-      BitNodeMultipliers.ServerStartingMoney = dec;
-      BitNodeMultipliers.ServerGrowthRate = dec;
-      BitNodeMultipliers.ServerWeakenRate = dec;
-
-      //Does not scale, otherwise security might start at 300+
-      BitNodeMultipliers.ServerStartingSecurity = 1.5;
-
-      BitNodeMultipliers.HomeComputerRamCost = inc;
-
-      BitNodeMultipliers.PurchasedServerCost = inc;
-      BitNodeMultipliers.PurchasedServerLimit = dec;
-      BitNodeMultipliers.PurchasedServerMaxRam = dec;
-      BitNodeMultipliers.PurchasedServerSoftcap = inc;
-
-      BitNodeMultipliers.ManualHackMoney = dec;
-      BitNodeMultipliers.ScriptHackMoney = dec;
-      BitNodeMultipliers.CompanyWorkMoney = dec;
-      BitNodeMultipliers.CrimeMoney = dec;
-      BitNodeMultipliers.HacknetNodeMoney = dec;
-      BitNodeMultipliers.CodingContractMoney = dec;
-
-      BitNodeMultipliers.CompanyWorkExpGain = dec;
-      BitNodeMultipliers.ClassGymExpGain = dec;
-      BitNodeMultipliers.FactionWorkExpGain = dec;
-      BitNodeMultipliers.HackExpGain = dec;
-      BitNodeMultipliers.CrimeExpGain = dec;
-
-      BitNodeMultipliers.FactionWorkRepGain = dec;
-      BitNodeMultipliers.FactionPassiveRepGain = dec;
-      BitNodeMultipliers.RepToDonateToFaction = inc;
-
-      BitNodeMultipliers.AugmentationRepCost = inc;
-      BitNodeMultipliers.AugmentationMoneyCost = inc;
-
-      BitNodeMultipliers.InfiltrationMoney = dec;
-      BitNodeMultipliers.InfiltrationRep = dec;
-
-      BitNodeMultipliers.FourSigmaMarketDataCost = inc;
-      BitNodeMultipliers.FourSigmaMarketDataApiCost = inc;
-
-      BitNodeMultipliers.CorporationValuation = dec;
-
-      BitNodeMultipliers.BladeburnerRank = dec;
-      BitNodeMultipliers.BladeburnerSkillCost = inc;
-
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = inc;
-      BitNodeMultipliers.StaneksGiftExtraSize = inc;
-      BitNodeMultipliers.GangSoftcap = 0.8;
-      BitNodeMultipliers.CorporationSoftCap = 0.8;
-      BitNodeMultipliers.WorldDaemonDifficulty = inc;
-
-      BitNodeMultipliers.GangUniqueAugs = dec;
-      break;
-    }
-    case 13: {
-      BitNodeMultipliers.PurchasedServerSoftcap = 1.6;
-
-      BitNodeMultipliers.HackingLevelMultiplier = 0.25;
-      BitNodeMultipliers.StrengthLevelMultiplier = 0.7;
-      BitNodeMultipliers.DefenseLevelMultiplier = 0.7;
-      BitNodeMultipliers.DexterityLevelMultiplier = 0.7;
-      BitNodeMultipliers.AgilityLevelMultiplier = 0.7;
-
-      BitNodeMultipliers.ServerMaxMoney = 0.45;
-      BitNodeMultipliers.ServerStartingMoney = 0.75;
-
-      BitNodeMultipliers.ServerStartingSecurity = 3;
-
-      BitNodeMultipliers.ScriptHackMoney = 0.2;
-      BitNodeMultipliers.CompanyWorkMoney = 0.4;
-      BitNodeMultipliers.CrimeMoney = 0.4;
-      BitNodeMultipliers.HacknetNodeMoney = 0.4;
-      BitNodeMultipliers.CodingContractMoney = 0.4;
-
-      BitNodeMultipliers.CompanyWorkExpGain = 0.5;
-      BitNodeMultipliers.ClassGymExpGain = 0.5;
-      BitNodeMultipliers.FactionWorkExpGain = 0.5;
-      BitNodeMultipliers.HackExpGain = 0.1;
-      BitNodeMultipliers.CrimeExpGain = 0.5;
-
-      BitNodeMultipliers.FactionWorkRepGain = 0.6;
-
-      BitNodeMultipliers.FourSigmaMarketDataCost = 10;
-      BitNodeMultipliers.FourSigmaMarketDataApiCost = 10;
-
-      BitNodeMultipliers.CorporationValuation = 0.001;
-
-      BitNodeMultipliers.BladeburnerRank = 0.45;
-      BitNodeMultipliers.BladeburnerSkillCost = 2;
-      BitNodeMultipliers.StaneksGiftPowerMultiplier = 2;
-      BitNodeMultipliers.StaneksGiftExtraSize = 1;
-      BitNodeMultipliers.GangSoftcap = 0.3;
-      BitNodeMultipliers.CorporationSoftCap = 0.3;
-      BitNodeMultipliers.WorldDaemonDifficulty = 3;
-      BitNodeMultipliers.GangUniqueAugs = 0.1;
-      break;
-    }
-    default:
-      console.warn("Player.bitNodeN invalid");
-      break;
-  }
-}
