+++ src/Gang/formulas/formulas.ts
@@ -1,4 +1,5 @@
-import { BitNodeMultipliers } from "../../BitNode/BitNodeMultipliers";
+import { Player } from "../../Player";
+import { getBitNodeMultipliers } from "../../BitNode/BitNodeMultipliers";
 import { GangMember } from "../GangMember";
 import { GangMemberTask } from "../GangMemberTask";
 
@@ -24,7 +25,7 @@ export function calculateRespectGain(gang: Gang, member: GangMember, task: GangM
   statWeight -= 4 * task.difficulty;
   if (statWeight <= 0) return 0;
   const territoryMult = Math.max(0.005, Math.pow(gang.territory * 100, task.territory.respect) / 100);
-  const territoryPenalty = (0.2 * gang.territory + 0.8) * BitNodeMultipliers.GangSoftcap;
+  const territoryPenalty = (0.2 * gang.territory + 0.8) * getBitNodeMultipliers(Player).GangSoftcap;
   if (isNaN(territoryMult) || territoryMult <= 0) return 0;
   const respectMult = calculateWantedPenalty(gang);
   return Math.pow(11 * task.baseRespect * statWeight * territoryMult * respectMult, territoryPenalty);
@@ -68,7 +69,7 @@ export function calculateMoneyGain(gang: Gang, member: GangMember, task: GangMem
   const territoryMult = Math.max(0.005, Math.pow(gang.territory * 100, task.territory.money) / 100);
   if (isNaN(territoryMult) || territoryMult <= 0) return 0;
   const respectMult = calculateWantedPenalty(gang);
-  const territoryPenalty = (0.2 * gang.territory + 0.8) * BitNodeMultipliers.GangSoftcap;
+  const territoryPenalty = (0.2 * gang.territory + 0.8) * getBitNodeMultipliers(Player).GangSoftcap;
   return Math.pow(5 * task.baseMoney * statWeight * territoryMult * respectMult, territoryPenalty);
 }
