+++ src/Netscript/RecentScripts.ts
@@ -1,4 +1,4 @@
-import { RunningScript } from "src/Script/RunningScript";
+import { RunningScript } from "../Script/RunningScript";
 import { Settings } from "../Settings/Settings";
 import { WorkerScript } from "./WorkerScript";
