+++ src/Programs/data/ProgramsMetadata.ts
@@ -8,7 +8,7 @@ import { IPlayer } from "../../PersonObjects/IPlayer";
 import { convertTimeMsToTimeElapsedString } from "../../utils/StringHelperFunctions";
 import { GetServer } from "../../Server/AllServers";
 import { numeralWrapper } from "../../ui/numeralFormat";
-import { BitNodeMultipliers } from "../../BitNode/BitNodeMultipliers";
+import { getBitNodeMultipliers } from "../../BitNode/BitNodeMultipliers";
 import { BitFlumeEvent } from "../../BitNode/ui/BitFlumeModal";
 import { calculateHackingTime, calculateGrowTime, calculateWeakenTime } from "../../Hacking";
 import { FactionNames } from "../../Faction/data/FactionNames";
@@ -307,7 +307,7 @@ export const programsMetadata: IProgramCreationParams[] = [
     name: "fl1ght.exe",
     create: null,
     run: (router: IRouter, terminal: ITerminal, player: IPlayer): void => {
-      const numAugReq = Math.round(BitNodeMultipliers.DaedalusAugsRequirement * 30);
+      const numAugReq = Math.round(getBitNodeMultipliers(player).DaedalusAugsRequirement * 30);
       const fulfilled = player.augmentations.length >= numAugReq && player.money > 1e11 && player.hacking >= 2500;
       if (!fulfilled) {
         terminal.print(`Augmentations: ${player.augmentations.length} / ${numAugReq}`);
