+++ src/Server/formulas/grow.ts
@@ -1,6 +1,6 @@
 import { CONSTANTS } from "../../Constants";
 import { Server } from "../Server";
-import { BitNodeMultipliers } from "../../BitNode/BitNodeMultipliers";
+import { getBitNodeMultipliers } from "../../BitNode/BitNodeMultipliers";
 import { IPlayer } from "../../PersonObjects/IPlayer";
 
 export function calculateServerGrowth(server: Server, threads: number, p: IPlayer, cores = 1): number {
@@ -16,7 +16,7 @@ export function calculateServerGrowth(server: Server, threads: number, p: IPlaye
   //Calculate adjusted server growth rate based on parameters
   const serverGrowthPercentage = server.serverGrowth / 100;
   const numServerGrowthCyclesAdjusted =
-    numServerGrowthCycles * serverGrowthPercentage * BitNodeMultipliers.ServerGrowthRate;
+    numServerGrowthCycles * serverGrowthPercentage * getBitNodeMultipliers(p).ServerGrowthRate;
 
   //Apply serverGrowth for the calculated number of growth cycles
   const coreBonus = 1 + (cores - 1) / 16;
