+++ src/StockMarket/StockMarketCosts.ts
@@ -1,12 +1,13 @@
-import { BitNodeMultipliers } from "../BitNode/BitNodeMultipliers";
+import { Player } from "../Player";
+import { getBitNodeMultipliers } from "../BitNode/BitNodeMultipliers";
 import { CONSTANTS } from "../Constants";
 
 export function getStockMarket4SDataCost(): number {
-  return CONSTANTS.MarketData4SCost * BitNodeMultipliers.FourSigmaMarketDataCost;
+  return CONSTANTS.MarketData4SCost * getBitNodeMultipliers(Player).FourSigmaMarketDataCost;
 }
 
 export function getStockMarket4STixApiCost(): number {
-  return CONSTANTS.MarketDataTixApi4SCost * BitNodeMultipliers.FourSigmaMarketDataApiCost;
+  return CONSTANTS.MarketDataTixApi4SCost * getBitNodeMultipliers(Player).FourSigmaMarketDataApiCost;
 }
 
 export function getStockMarketWseCost(): number {
